Fixed comparison warning.

+++ dev/Source/DynamicStackArray.hpp
@@ -7,7 +7,7 @@
  */
 #define LOTTERY_DYNAMIC_STACK_ARRAY(TYPE, NAME, SIZE, DEFAULT_VALUE)\
     TYPE *NAME = (TYPE *)alloca(sizeof(TYPE) * SIZE);\
-    for (size_t i = 0; i < SIZE; ++i) { NAME[i] = DEFAULT_VALUE; }
+    for (size_t i = 0; i < (size_t)SIZE; ++i) { NAME[i] = DEFAULT_VALUE; }
 
 
 #endif //LOTTERY_DYNAMIC_STACK_ARRAY_HPP
